+++ app/src/main/java/com/plusorminustwo/postmark/di/DatabaseModule.kt
@@ -26,7 +26,8 @@ object DatabaseModule {
                 PostmarkDatabase.MIGRATION_4_5,
                 PostmarkDatabase.MIGRATION_5_6,
                 PostmarkDatabase.MIGRATION_6_7,
-                PostmarkDatabase.MIGRATION_7_8
+                PostmarkDatabase.MIGRATION_7_8,
+                PostmarkDatabase.MIGRATION_8_9
             )
             .addCallback(PostmarkDatabase.FTS_CALLBACK)
             .build()